Feat: 상품 클릭 후 2초동안 재상품 선택 및 투입, 반환 막기
closed issue #16
